receiver/requirements.txt
aiogram
python-dotenv
notify-send
aiotdlib